forward signals to child
Before, trying to kill mtimeout with Ctrl-C or the `kill` command would
not really work. The <cmd> it was running would keep running.

Now, it will forward most signals to the child, usually causing the
child to exit in the above case.

Also added test/signals.sh.
Not a full test suite, just a "looks right" sorta script.